Added Flex Box
Made 3 boxes that staked ontop of each other but want them to be side-by-side. Class "cards" is the parent and the class "card1, card2, card3" are the child boxes (child will act as parent wants). Want the child boxes to be flex boxes to create the side-by-side layout; therefore, in .css file = under cards = added "display: flex."