Learned 'Em' and 'Rem' and changed to rem for proper website
Em and Rem for website are recommened for to accommodate visually impaired and to keep consistent with other web browsers. To calculate rem or em - can use rem/em calculators online. Rem = (font size desired) x 0.0625. You would input into this instead of the 'px' size. Be informed about the diff. between em and rem. "Fontcraziness" is an exercise that encompasses the em and rem learning with Daniel Scott.