Making a DIV container clickable
Added <a> to each child "class=card" instead of the parent of "cards." Inserting <a> under per class=card will make the container clickable without affecting the container itself (ie. box sizes). If you put the <a herf> under the parent "card", it will affect the correlation between the parent and child; therefore, not distorting the container (ie. narrows it). So you will have to put it with the child tags and after <div class="card.." >