Created clickable button
Added an 'a' tage and gave it a class of "button". Created a test-page.html to test if the link works when the "Book Online" button is pressed. Note: place the button in the section you want it to be and to be after/before. Then added specification for button in the css file. Will add proper padding for it in the next module.