Properly added margin padding to clickable Book Online button
Learned about inline, block and how it makes a difference if want to add padding around the source. Block = it's own line and pushes each other to their own line. Inline = stays within the same line as another, does not push each other to the next line. So if want to create padding around the "context" - will need to make the button be a display: inline-block = has a bit of both where you want the button and text on the same line but have a padding around the button. Then you can add the padding with margin (top or bottom, etc) in the  .button css or have an "a" tag with button (a.button). Either way can work depends what is being coded and what will be affected.