Box/Text Shadows, Border-Radius
Added border radius to button and drop-shadows for the button, H1 text and 3 images (card1, card2, card3). Note: BOX-SHADOW=for creating drop shadows for "box" objects. TEXT-SHADOW=to create drop shadow for texts. Will need to add a color of drop shadow (for it to be visible) at the end of the text/box-shadow code in css.