Flex-wrap and adding line in header section
Added a line on the header to create the header/body division. Also, adjusted the padding around the top and bottom the line.

Learned about flex-wrap. Created a testimonials at the bottom of the page and used flex-wrap to properly layout this section under the class=testimonials. Added specifica styles for this box containing the text, and then, specified the styles for .tbox class.

Removed the html height: 100% so the background gradient can extend more longer and becomes continuous, especially when adding more content below the min height. Add padding/margin at the bottom of the testimonials (.tbox), NOT ON THE BODY tag because it might mess up the body part later on.