Learned Psuedo Selector
Added the style for pseudo class of last-child for the name in the testimonials. Used specific .tbox p where the 'p' is specifically targeting the p tags in tboxes to make it the 'last-child'. This type of specific detail is good because if you just have it as .tbox without specifying the p, it can target the content you put AFTER the testimonials for that style if that isn't what you want.