Link for Contact Page
Created a Contact Us page and linked it to the index.html page. Linked the pages to be able to go back and forth for the Contact Us and Home page (when you click on the lion logo). Input the content for the Contact Us page and stylized it with top padding and min-height of 300px. Gave this a div class of main2 instead of using <main> as <main> is an actual html tag.