Simple PHP Form
Added mail.php form to link with contact us (did not check if working due to not having a server to host it). However, learned to add  the code for it. Also, input placeholder text and labels to website form text fields in HTML. Created margin-bottom spacing between the fields.