Message/text box, check mart tick box and radio button with dot in middle on HTML form
Created a text box for user to send a message to website creator. Also, added a check mark box for email subscription (which can be checked and unchecked), and radio buttons to give an option (where one OR the other gets ticked. Once ticked, it cannot be unticked and therefore, having an option for another selection is better to have).