Style form text boxes and check boxes
Added margin/padding between inline boxes, make the text for the first name and email to display as block, not inline as want them on separate lines. Stylize the 'fire away' button by giving it a class of button. Added a class of "margin-right-a" to make spaces between the text and fields (ie. radio button and text, dropdown menu and datetimeline and color section).